Licensing/LicensingSystem/requirements.txt
alembic==1.5.5
APScheduler==3.7.0
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
h5py==3.1.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.3
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pygame==2.0.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
psycopg2==2.8.6